backend/requirements.txt
fastapi==0.109.0
gunicorn==21.2.0
uvicorn==0.27.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
email-validator==2.1.0.post1
firebase-admin==6.4.0
websockets>=12.0
